If Russ Ackoff had given a TED Talk...
... it would have been great. Watch it here:
https://www.youtube.com/watch?v=OqEeIG8aPPk

This is a great talk about Systems Thinking: "A system has properties
that each of its parts doesn't have. Think of an automobile, it can
transport you from A to B, but none of its parts can. A motor can't even
transport itself. Yet if you take away one of its parts, it doesn't work
anymore."