Read "The Circle" while I was on vacation
It's a fine book and it sketches out some really dystopian views about a
future where a company knows everything about everyone, from an employer
point of view, and from an all-seeing-eye
knows-everything-about-everyone point of view. It clearly tries to
balance the "good" vs the "bad" and the "ugly", but the story suffers a
bit due to the super descriptive style of writing. You'll be a lot of
pages in before something interesting starts happening. And then that
part could have been more elaborate...

Good book overall, but not the best i've read.