Listened to Rich Hickey complain about Haskell
Instead of telling listeners how Clojure helps you, he's bragging about
how it's not like statically typed languages in any way, and complaining
about how difficult programming using static types is for him.

He reasons that Types are not Semantics. I disagree. Semantics of a
system describe the relationship from the input to the output of that
system, which is a strong part of the type system. Making invalid state
of a system unrepresentable.

Of course, he's got good points too, like the fact that concurrency is
easy if you got immutability, and functional programming without side
effects.

In short, i think he's a great speaker, but his opinion is
short-sighted, i bet he hasn't got a lot of experience with strongly
typed functional programming languages, and he's really only seeing the
things he wants to see. Of course Clojure is a great language, after
all, it's a lisp and HE WROTE IT. i get it.

https://www.youtube.com/watch?v=2V1FtfBDsLU