Watched this great talk about organisation boundaries from last DDDEU
Rewatching @ntcoding 's talk from @ddd_eu last year, the "Autonomy
Context" and the great insight in how people & teams within companies
can work together in a smarter way still resonate with me in a glorious
way.

I strongly believe in the fact that companies should be more agile in
the original sense of the word, which as a result means that there need
to be defined boundaries within which teams can operate in a very agile
way. I think this talk demonstrates that.

Enjoyed it at the conference, and I enjoyed it even more rewatching on
the couch at home with a beer. Great speaker, good insights, right
mindset.

https://www.youtube.com/watch?v=5bgxzaHCyCs