Listened to @mathiasverraes talk about Emergent Boundaries
https://www.youtube.com/watch?v=ECM1rPYxvD4

I really enjoy the fact that Mathias & some others have done more and
more presenting without slides, and using alternative methods for
visualising stuff while they're talking. In this talk you can see
Mathias use his trusty stickies & a whiteboard for that.

The talk itself is consise and to the point, and contained some nice
a-ha moments for some of the typical DDD concepts & misconceptions out
there, e.g. (in the beginning of the talk) the bounded contexts that
aren't technical or operational contexts but domain contexts, which
people often don't get.

overall, great talk.