Listened to this keynote about versioning and backwards compatibility
https://www.youtube.com/watch?v=oyLBGkS5ICk

Rich Hickey's doing a rant on semantic versioning and API breaks, but he
has a great point: backwards compatibility is very important.
Immutability is important and can help. I don't completely agree with
the fact that we have to keep all old API's intact in all kinds of
systems. If you didn't grasp a concept when you first created a system,
and threw it away and replaced it with something that fits better, the
old part of the system shouldn't keep working as far as i'm concerned.

Again, from his point of view, this talk makes sense, but i'm convinced
that it's not necessarily true for everyone. Semantic versioning is
there for you if you *need* to do BC breaks. The problem with the two
packages that have two times the same dependencies, but other versions,
is something that can be solved by using some techniques explained very
nicely by Matthias Noback in "principles of package design".

Anyway, good talk, got some things out of it. We certainly need to think
more about breaking changes...