Read the latest book by Matthias Noback
Although I think it's a great handbook with lots of hands-on information
about working with a microservices based architecture, I was not
completely blown away by the book. Some information seems redundant,
there's lots of docker info (which doesn't attract my interest too much)
and there are big parts that I already "felt" because of reading Eric
Evans and Vaughn Vernon Domain Driven Design books. Overall a great book
for people wanting to get to know microservices and playing around with
the concepts associated with them.