Read these two amazing sci-fi books will travelling!
I already read Brave New World in the past but it was great to refresh
my mind about it. I had totally forgotten the plot, but i had remembered
the world created in the book.

The War Of The Worlds, my girlfriend brought me from the library and
it's a fantastic book. It's so old, but still relevant. The technology
described is cool. The story is nice and sound. It's told from a "me, I"
first person perspective and feels very realistic. Loved it!