Watched two great DDDEU talks
https://www.youtube.com/watch?v=v0pNA8a7dbY
https://www.youtube.com/watch?v=hyJ7LJPhIWI

Really liked the second one, where talks about teams:

"If you want people to work in a group, you need to treat them as a
group, not as separate individuals, and yet that's we do... How do we
pay them? How do we do assessments? How do we interview for them? Is
everyone on their own laptop?"