Watched this great talk by Bernd Ruecker
"Lost in transaction" by @berndruecker is about the typical "one
transaction for multiple aggregates" problem, and comes with tons of
knowledge about how to tackle it 😎

I love the talk because it starts of with mapping out the problem and
then looking for possible solutions and explaining the building blocks
for them, e.g. Eventual Consistency & Idempotency to be able to chose
Orchestration or Choreography between services

Using Saga "pattern" to have undo's or compensating actions for certain
tasks, etc. Finally talks about BPMN (google it if you don't know it),
DSL's and lots of other good stuff. Definitely a good talk for those who
want some technical <hands-on> knowledge! 👏 👨‍🏭

https://www.youtube.com/watch?v=WRR26jJNh68