Watched another great DDDEU talk about Empathy
"Exploring Empathy With Algorithms" by @andreagoulet Empathy is a skill
that we can learn, one of the basic tools to improve communication and
e.g. take away knowledge silos. I found the "communication artifacts"
and "6 types of empathy" very interesting 😇

It's also very nice to see someone who's "a business person", not an
engineer or developer, bringing up examples of situations that are
tailored to the more technical oriented minds. Empathy, you know.  🙌 🖖

https://www.youtube.com/watch?v=rvUsLaCFu-s