Went to watch the new Star Wars yesterday...
I liked it and didn't like it at the same time... It's too long, too
much stuff is crammed in there, and it's at times too much Disney™
style... Also too much of the same in one movie: going to a planetary
system, being followed by Kylo Ren, fight a little, move to the next
system, restart the sequence.

On the other hand, it's nice to conclude this chapter in a quite decent
way, story wise.