Watched Casa De Papel!
This must 've been one of my favorite series of the past years, together
with the handmaid's tale. I love how it's filmed, how there's a plan for
everything, but that some things still don't work out as expected. Good
amounts of tension and humor as well. Gotta love it!