Forgot to record that we went to Gentjazz 1.5
The COVID-19 edition of Gentjazz festival was the first one in over 10
years that I didn't work there as a volunteer. We went there with Bram,
Haydé, Jan, Aaron and Dieter Beke. Nice evening, although the nicest
part was the company, and being with people, rather than the concerts
(which were also nice).